+++ crawl-ref/source/command.cc
@@ -332,7 +332,9 @@ void list_jewellery()
                  (i == EQ_RING_SIX)    ? "6th ring" :
                  (i == EQ_RING_SEVEN)  ? "7th ring" :
                  (i == EQ_RING_EIGHT)  ? "8th ring" :
-                 (i == EQ_RING_AMULET) ? "Amulet ring"
+                 (i == EQ_RING_AMULET) ? "Amulet ring" :
+                 (i == EQ_AMULET_LEFT) ? "Left Amulet" :
+                 (i == EQ_AMULET_RIGHT) ? "Right Amulet"
                                        : "unknown";
 
         string item;