+++ crawl-ref/source/player-equip.cc
@@ -136,16 +136,35 @@ static void _assert_valid_slot(equipment_type eq, equipment_type slot)
 #ifdef ASSERTS
     if (eq == slot)
         return;
-    ASSERT(eq == EQ_RINGS); // all other slots are unique
-    equipment_type r1 = EQ_LEFT_RING, r2 = EQ_RIGHT_RING;
-    if (you.species == SP_OCTOPODE)
-        r1 = EQ_RING_ONE, r2 = EQ_RING_EIGHT;
-    if (slot >= r1 && slot <= r2)
-        return;
-    if (const item_def* amu = you.slot_item(EQ_AMULET, true))
-        if (is_unrandom_artefact(*amu, UNRAND_FINGER_AMULET) && slot == EQ_RING_AMULET)
+    ASSERT(eq == EQ_WEAPON || eq == EQ_RINGS || eq == EQ_AMULETS); // all other slots are unique
+    if (eq == EQ_WEAPON) {
+        if(slot != EQ_SECOND_WEAPON) {
+            die("weapon on invalid slot %d", slot);
+        }
+    }
+    if (eq == EQ_RINGS) {
+        equipment_type r1 = EQ_LEFT_RING, r2 = EQ_RIGHT_RING;
+        if (you.species == SP_OCTOPODE)
+            r1 = EQ_RING_ONE, r2 = EQ_RING_EIGHT;
+        if (slot >= r1 && slot <= r2)
             return;
-    die("ring on invalid slot %d", slot);
+
+        if (const item_def * amu = you.slot_item(EQ_AMULET, true))
+            if (is_unrandom_artefact(*amu, UNRAND_FINGER_AMULET) && slot == EQ_RING_AMULET)
+                return;
+        die("ring on invalid slot %d", slot);
+    }
+    else if(eq == EQ_AMULETS) {
+        if (you.species == SP_TWO_HEADED_OGRE) {
+            if (slot >= EQ_AMULET_LEFT && slot <= EQ_AMULET_RIGHT) {
+                return;
+            }
+        }
+        else if (slot == EQ_AMULET) {
+            return;
+        }
+        die("amulet on invalid slot %d", slot);
+    }
 #endif
 }
 
@@ -154,7 +173,8 @@ void equip_effect(equipment_type slot, int item_slot, bool unmeld, bool msg)
     item_def& item = you.inv[item_slot];
     equipment_type eq = get_item_slot(item);
 
-    if (slot == EQ_WEAPON && eq != EQ_WEAPON)
+    if ((slot == EQ_WEAPON && eq != EQ_WEAPON) || 
+        (slot == EQ_SECOND_WEAPON && eq != EQ_WEAPON))
         return;
 
     _assert_valid_slot(eq, slot);
@@ -164,7 +184,7 @@ void equip_effect(equipment_type slot, int item_slot, bool unmeld, bool msg)
 
     const interrupt_block block_unmeld_interrupts(unmeld);
 
-    if (slot == EQ_WEAPON)
+    if (slot == EQ_WEAPON || slot == EQ_SECOND_WEAPON)
         _equip_weapon_effect(item, msg, unmeld);
     else if (slot >= EQ_CLOAK && slot <= EQ_BODY_ARMOUR)
         _equip_armour_effect(item, unmeld, slot);
@@ -177,14 +197,15 @@ void unequip_effect(equipment_type slot, int item_slot, bool meld, bool msg)
     item_def& item = you.inv[item_slot];
     equipment_type eq = get_item_slot(item);
 
-    if (slot == EQ_WEAPON && eq != EQ_WEAPON)
+    if ((slot == EQ_WEAPON && eq != EQ_WEAPON) || 
+        (slot == EQ_SECOND_WEAPON && eq != EQ_WEAPON))
         return;
 
     _assert_valid_slot(eq, slot);
 
     const interrupt_block block_meld_interrupts(meld);
 
-    if (slot == EQ_WEAPON)
+    if (slot == EQ_WEAPON || slot == EQ_SECOND_WEAPON)
         _unequip_weapon_effect(item, msg, meld);
     else if (slot >= EQ_CLOAK && slot <= EQ_BODY_ARMOUR)
         _unequip_armour_effect(item, meld, slot);