+++ crawl-ref/source/transform.cc
@@ -44,7 +44,7 @@
 
 static const int EQF_NONE = 0;
 // "hand" slots (not rings)
-static const int EQF_HANDS = SLOTF(EQ_WEAPON) | SLOTF(EQ_SHIELD)
+static const int EQF_HANDS = SLOTF(EQ_WEAPON) | SLOTF(EQ_SECOND_WEAPON) | SLOTF(EQ_SHIELD)
                              | SLOTF(EQ_GLOVES);
 // core body slots (statue form)
 static const int EQF_STATUE = SLOTF(EQ_GLOVES) | SLOTF(EQ_BOOTS)
@@ -62,7 +62,7 @@ static const int EQF_RINGS = SLOTF(EQ_LEFT_RING) | SLOTF(EQ_RIGHT_RING)
                              | SLOTF(EQ_RING_FIVE) | SLOTF(EQ_RING_SIX)
                              | SLOTF(EQ_RING_SEVEN) | SLOTF(EQ_RING_EIGHT);
 // amulet & pal
-static const int EQF_AMULETS = SLOTF(EQ_AMULET) | SLOTF(EQ_RING_AMULET);
+static const int EQF_AMULETS = SLOTF(EQ_AMULET) | SLOTF(EQ_AMULET_LEFT) | SLOTF(EQ_AMULET_RIGHT) | SLOTF(EQ_RING_AMULET);
 // everything
 static const int EQF_ALL = EQF_PHYSICAL | EQF_RINGS | EQF_AMULETS;
 
@@ -148,7 +148,7 @@ bool Form::can_wear_item(const item_def& item) const
     if (item.base_type == OBJ_JEWELLERY)
     {
         if (jewellery_is_amulet(item))
-            return slot_available(EQ_AMULET);
+            return !all_blocked(EQF_AMULETS);
         return !all_blocked(EQF_RINGS);
     }